+++ packages/next-example/src/app/sign/page.tsx
@@ -11,6 +11,7 @@ import {
   TSignUpContinueHandler,
   SignUpValue,
   TModalMethodRef,
+  setLoading,
 } from '@portkey/did-ui-react';
 import { ChainId } from '@portkey/types';
 import { sleep } from '@portkey/utils';
@@ -147,7 +148,7 @@ export default function Sign() {
         termsOfService={'https://portkey.finance/terms-of-service'}
         privacyPolicy={'https://portkey.finance/privacy-policy'}
         onFinish={async res => {
-          console.log(res, 'onFinish====');
+          console.log(res, 'onFinish====', did.didWallet);
           CHAIN_ID = res.chainId;
           did.save(PIN);
         }}
@@ -158,8 +159,16 @@ export default function Sign() {
           ref.current?.setOpen(false);
           setLifeCycle(undefined);
         }}
-        onCreatePending={info => {
+        onCreatePending={async info => {
           console.log(info, 'onCreatePending====info');
+          CHAIN_ID = info.didWallet?.chainId || 'AELF';
+          did.save(PIN);
+          ref.current?.setOpen(false);
+          setLoading(false);
+          console.log('did.DIDWallet==onCreatePending', did.didWallet);
+          localStorage.setItem('sessionId', JSON.stringify({ sessionId: info.sessionId }));
+          // const wallet = await did.load(PIN);
+          // console.log(wallet, did.didWallet, 'onCreatePending===wallet==load');
         }}
         onSignUp={onSignUpHandler}
         // defaultLifeCycle={{ LoginByScan: null }}