+++ packages/types/src/wallet.ts
@@ -108,6 +108,11 @@ export interface IReferralInfo {
   projectCode?: string;
 }
 
+export enum LoginStatusEnum {
+  INIT = 'INIT',
+  SUCCESS = 'SUCCESS',
+  FAIL = 'FAIL',
+}
 export interface IExtraInfo {
   ip?: string;
   activityId?: string;